# Speed up ring finding by skipping nodes not in rings