When finding rings - If we've exhaustively searched
an atom and found no rings, we should mark the bonds to
that atom as "not in a ring". We can also mark any
neighboring low degree atoms in the same way.

This speeds up searches in large molecules, because
a ring search that _doesn't_ find any rings is very
expensive, and it's a bummer to pay for that search
on two neighboring atoms, for instance.

Ring finding is a bottleneck in sanitization,
especially for larger molecules.

Here are some SSSR timings on protiens from the RCSB PDB:

Before this commit:

 - 3EOH: 3.5s
 - 2J3N: 1.2s
 - 1NKS: 0.05s
    
After this commit:

-  3EOH: 0.8s
-  2J3N: 0.3s
-  1NKS: 0.02s